stores as interfaces
Fix #18
