+++ src/store/memo_trustanchorstore.rs
@@ -1,74 +1,27 @@
 // Copyright 2023 Contributors to the Veraison project.
 // SPDX-License-Identifier: Apache-2.0
 
+use super::cpak::Cpak;
 use super::errors::Error;
-use hex_literal::hex;
-use jsonwebkey as jwk;
-use jwk::JsonWebKey;
-use serde::Deserialize;
-use serde_json::value::RawValue;
+use super::ITrustAnchorStore;
+use jsonwebkey::JsonWebKey;
 use std::collections::HashMap;
 use std::sync::RwLock;
-
-/// A CCA platform attestation key and associated metadata
-#[serde_with::serde_as]
-#[derive(Clone, Deserialize, Debug)]
-pub struct Cpak {
-    /// The CPAK (a raw public key) wrapped in a Subject Public Key Info and
-    /// serialised using the textual encoding described in §13 of RFC7468
-    #[serde(rename(deserialize = "pkey"))]
-    raw_pkey: Box<RawValue>,
-
-    #[serde(skip)]
-    pkey: Option<jwk::JsonWebKey>,
-
-    /// The CCA platform Implementation ID claim uniquely identifies the
-    /// implementation of the CCA platform.  The semantics of the CCA platform
-    /// Implementation ID value are defined by the manufacturer or a particular
-    /// certification scheme.  For example, the ID could take the form of a
-    /// product serial number, database ID, or other appropriate identifier.
-    /// It is a fixed-size, 32 bytes binary blob, base64 encoded.
-    #[serde(rename(deserialize = "implementation-id"))]
-    #[serde_as(as = "serde_with::hex::Hex")]
-    impl_id: [u8; 32],
-
-    /// The CCA platform Instance ID claim represents the unique identifier of
-    /// the CPAK
-    /// It is a fixed-size, 33 bytes binary blob (a EAT UEID), base64 encoded.
-    /// The first byte MUST be 0x01 (i.e., RAND UEID).
-    #[serde(rename(deserialize = "instance-id"))]
-    #[serde_as(as = "serde_with::hex::Hex")]
-    inst_id: [u8; 33],
-}
-
-impl Cpak {
-    pub fn parse_pkey(&mut self) -> Result<(), Error> {
-        let s = self.raw_pkey.get();
-
-        let pkey = serde_json::from_str::<JsonWebKey>(s)
-            .map_err(|e| Error::Syntax(e.to_string()))?
-            .clone();
-
-        self.pkey = Some(pkey);
-
-        Ok(())
-    }
-}
 
 /// The store where the active CPAKs are stashed.  CPAKs are indexed by their
 /// instance-id.
 #[derive(Debug)]
-pub struct TrustAnchorStore {
+pub struct MemoTrustAnchorStore {
     p: RwLock<HashMap<[u8; 33], Cpak>>,
 }
 
-impl Default for TrustAnchorStore {
+impl Default for MemoTrustAnchorStore {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl TrustAnchorStore {
+impl MemoTrustAnchorStore {
     /// Returns a new empty TrustAnchorStore
     pub fn new() -> Self {
         Self {
@@ -89,9 +42,11 @@ impl TrustAnchorStore {
 
         Ok(())
     }
+}
 
+impl ITrustAnchorStore for MemoTrustAnchorStore {
     /// Lookup a trust anchor from the store given the corresponding Instance ID
-    pub fn lookup(&self, inst_id: &[u8; 33]) -> Option<Cpak> {
+    fn lookup(&self, inst_id: &[u8; 33]) -> Option<Cpak> {
         return self.p.read().unwrap().get(inst_id).cloned();
     }
 }
@@ -106,7 +61,7 @@ mod tests {
 
     #[test]
     fn load_json_and_lookup_ok() {
-        let mut s: TrustAnchorStore = Default::default();
+        let mut s: MemoTrustAnchorStore = Default::default();
 
         // load store from JSON
         s.load_json(TEST_JSON_TA_OK_0).unwrap();
