+++ src/store/mod.rs
@@ -1,12 +1,20 @@
 // Copyright 2023 Contributors to the Veraison project.
 // SPDX-License-Identifier: Apache-2.0
 
+pub use self::cpak::Cpak;
 pub use self::errors::Error;
-pub use self::refval::PlatformRefValue;
-pub use self::refval::RealmRefValue;
-pub use self::refval::RefValueStore;
-pub use self::trustanchor::TrustAnchorStore;
+pub use self::irefvaluestore::IRefValueStore;
+pub use self::itrustanchorstore::ITrustAnchorStore;
+pub use self::memo_refvaluestore::MemoRefValueStore;
+pub use self::memo_trustanchorstore::MemoTrustAnchorStore;
+pub use self::platformrefvalue::PlatformRefValue;
+pub use self::realmrefvalue::RealmRefValue;
 
+mod cpak;
 mod errors;
-mod refval;
-mod trustanchor;
+mod irefvaluestore;
+mod itrustanchorstore;
+mod memo_refvaluestore;
+mod memo_trustanchorstore;
+mod platformrefvalue;
+mod realmrefvalue;
